% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbcon.R
\name{getDecryptParam}
\alias{getDecryptParam}
\title{Initialize db connection object}
\usage{
getDecryptParam(str, encrypted, key)
}
\arguments{
\item{str}{string to decrypt}

\item{encrypted}{T or F}

\item{key}{secret key}
}
\description{
This package uses yaml file to connect/close db using necessary packages.
Params in yaml maybe encypted using safer package
}
\details{
Supported databases:  SQL and MySQL
Used for decrypting connection params
}
