% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbcon.R
\name{init}
\alias{init}
\title{Initialize db connection object}
\usage{
init(file, db_object)
}
\arguments{
\item{file}{Path to db config => valid yaml file}

\item{db_object}{database object you want to connect to}
}
\value{
A connection object as con
}
\description{
This function uses yaml file to connect to db using necessary packages.
Supported databases:  SQL and MySQL
Below is the structure of expected object in yaml
db_name:
dataconnection:
driver: 'ODBC Driver 17 for SQL Server'
server: 'xx.xx.xx.xx'
uid: 'xxx'
pwd: 'xx'
port: 1433
db_type: 'SQL'  or MySQL
database: 'xxx'
}
