% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbcon.R
\name{openDb}
\alias{openDb}
\title{Open connection}
\usage{
openDb(file, db_object, encrypted_params = F, key = "")
}
\arguments{
\item{file}{Path to db config => valid yaml file}

\item{db_object}{database object you want to connect to}

\item{encrypted_params}{T if soap params are encrypted otherwise F}

\item{key}{Secret key for encryption}
}
\value{
A connection object as con
}
\description{
Open connection
}
